Remove gunicorn and start uvicorn directly
This app is mostly run in a container in k8s.
Running it with the process manager of gunicorn and a single uvicorn worker is not only unnecessary, but also makes signal handling slightly worse (e.g. SIGINT directly terminates uvicorn, but takes much longer in gunicorn).
It can also lead to a semi-healthy state, where the uvicorn worker is crashing, but the gunicorn process just spawns a new one, without surfacing the issue.

See also https://fastapi.tiangolo.com/deployment/docker/#replication-number-of-processes