Exit hard if one of the started tasks fails
When an exception occurs within an asyncio task, this doesn't interrupt the main flow.
This means that exceptions never surfaced and the webserver continued to serve metrics, although they might not have been updated anymore.

This wrapper catches any exception, logs it and then calls `sys.exit(1)`, so that the orchestrator (e.g. k8s) can recycle the container, allowing observability tools to notice that there is something wrong.